python .gitignore included frontend/lib directory which broke frontend on commit - fixed here by moving everything into utils directory
