+++ frontend/src/components/LineChart.tsx
@@ -28,7 +28,7 @@ import {
   type AvailableChartColorsKeys,
 } from "@/lib/chartUtils";
 import { useOnWindowResize } from "@/hooks/useOnWindowsResize";
-import { cx } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 
 //#region Legend
 
@@ -48,7 +48,7 @@ const LegendItem = ({
   const hasOnValueChange = !!onClick;
   return (
     <li
-      className={cx(
+      className={cn(
         // base
         "group inline-flex flex-nowrap items-center gap-1.5 rounded-sm px-2 py-1 whitespace-nowrap transition",
         hasOnValueChange
@@ -61,7 +61,7 @@ const LegendItem = ({
       }}
     >
       <span
-        className={cx(
+        className={cn(
           "h-[3px] w-3.5 shrink-0 rounded-full",
           getColorClassName(color, "bg"),
           activeLegend && activeLegend !== name ? "opacity-40" : "opacity-100"
@@ -69,7 +69,7 @@ const LegendItem = ({
         aria-hidden={true}
       />
       <p
-        className={cx(
+        className={cn(
           // base
           "truncate text-xs whitespace-nowrap",
           // text color
@@ -117,7 +117,7 @@ const ScrollButton = ({ icon, onClick, disabled }: ScrollButtonProps) => {
   return (
     <button
       type="button"
-      className={cx(
+      className={cn(
         // base
         "group inline-flex size-5 items-center truncate rounded-sm transition",
         disabled
@@ -254,13 +254,13 @@ const Legend = React.forwardRef<HTMLOListElement, LegendProps>((props, ref) => {
   return (
     <ol
       ref={ref}
-      className={cx("relative overflow-hidden", className)}
+      className={cn("relative overflow-hidden", className)}
       {...other}
     >
       <div
         ref={scrollableRef}
         tabIndex={0}
-        className={cx(
+        className={cn(
           "flex h-full",
           enableLegendSlider
             ? hasScroll?.right || hasScroll?.left
@@ -282,7 +282,7 @@ const Legend = React.forwardRef<HTMLOListElement, LegendProps>((props, ref) => {
       {enableLegendSlider && (hasScroll?.right || hasScroll?.left) ? (
         <>
           <div
-            className={cx(
+            className={cn(
               // base
               "absolute top-0 right-0 bottom-0 flex h-full items-center justify-center pr-1",
               // background color
@@ -341,7 +341,7 @@ const ChartLegend = (
     <div
       ref={legendRef}
       style={{ paddingLeft: paddingLeft }}
-      className={cx(
+      className={cn(
         "flex items-center",
         { "justify-center": legendPosition === "center" },
         { "justify-start": legendPosition === "left" },
@@ -391,7 +391,7 @@ const ChartTooltip = ({
     const legendPayload = payload.filter((item: any) => item.type !== "none");
     return (
       <div
-        className={cx(
+        className={cn(
           // base
           "rounded-md border text-sm shadow-md",
           // border color
@@ -400,9 +400,9 @@ const ChartTooltip = ({
           "bg-white dark:bg-gray-950"
         )}
       >
-        <div className={cx("border-b border-inherit px-4 py-2")}>
+        <div className={cn("border-b border-inherit px-4 py-2")}>
           <p
-            className={cx(
+            className={cn(
               // base
               "font-medium",
               // text color
@@ -412,7 +412,7 @@ const ChartTooltip = ({
             {label}
           </p>
         </div>
-        <div className={cx("space-y-1 px-4 py-2")}>
+        <div className={cn("space-y-1 px-4 py-2")}>
           {legendPayload.map(({ value, category, color }, index) => (
             <div
               key={`id-${index}`}
@@ -421,13 +421,13 @@ const ChartTooltip = ({
               <div className="flex items-center space-x-2">
                 <span
                   aria-hidden="true"
-                  className={cx(
+                  className={cn(
                     "h-[3px] w-3.5 shrink-0 rounded-full",
                     getColorClassName(color, "bg")
                   )}
                 />
                 <p
-                  className={cx(
+                  className={cn(
                     // base
                     "text-right whitespace-nowrap",
                     // text color
@@ -438,7 +438,7 @@ const ChartTooltip = ({
                 </p>
               </div>
               <p
-                className={cx(
+                className={cn(
                   // base
                   "text-right font-medium whitespace-nowrap tabular-nums",
                   // text color
@@ -600,7 +600,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
     return (
       <div
         ref={ref}
-        className={cx("h-80 w-full", className)}
+        className={cn("h-80 w-full", className)}
         tremor-id="tremor-raw"
         {...other}
       >
@@ -625,7 +625,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
           >
             {showGridLines ? (
               <CartesianGrid
-                className={cx("stroke-gray-200 stroke-1 dark:stroke-gray-800")}
+                className={cn("stroke-gray-200 stroke-1 dark:stroke-gray-800")}
                 horizontal={true}
                 vertical={false}
               />
@@ -643,7 +643,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
               }
               fill=""
               stroke=""
-              className={cx(
+              className={cn(
                 // base
                 "text-xs",
                 // text fill
@@ -673,7 +673,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
               tick={{ transform: "translate(-3, 0)" }}
               fill=""
               stroke=""
-              className={cx(
+              className={cn(
                 // base
                 "text-xs",
                 // text fill
@@ -767,7 +767,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
             ) : null}
             {categories.map((category) => (
               <Line
-                className={cx(
+                className={cn(
                   getColorClassName(
                     categoryColors.get(category) as AvailableChartColorsKeys,
                     "stroke"
@@ -790,7 +790,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
                   } = props;
                   return (
                     <Dot
-                      className={cx(
+                      className={cn(
                         "stroke-white dark:stroke-gray-950",
                         onValueChange ? "cursor-pointer" : "",
                         getColorClassName(
@@ -844,7 +844,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
                         strokeLinecap={strokeLinecap}
                         strokeLinejoin={strokeLinejoin}
                         strokeWidth={strokeWidth}
-                        className={cx(
+                        className={cn(
                           "stroke-white dark:stroke-gray-950",
                           onValueChange ? "cursor-pointer" : "",
                           getColorClassName(
@@ -875,7 +875,7 @@ const LineChart = React.forwardRef<HTMLDivElement, LineChartProps>(
             {onValueChange
               ? categories.map((category) => (
                   <Line
-                    className={cx("cursor-pointer")}
+                    className={cn("cursor-pointer")}
                     strokeOpacity={0}
                     key={category}
                     name={category}
