+++ frontend/src/components/ui/separator.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as SeparatorPrimitive from "@radix-ui/react-separator";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 
 function Separator({
   className,
