+++ frontend/src/components/ui/skeleton.tsx
@@ -1,4 +1,5 @@
-import { cn } from "~/lib/utils"
+import React from "react";
+import { cn } from "~/utils/utils";
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
@@ -7,7 +8,7 @@ function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
       className={cn("bg-accent animate-pulse rounded-md", className)}
       {...props}
     />
-  )
+  );
 }
 
-export { Skeleton }
+export { Skeleton };
