+++ frontend/src/utils/utils.ts
@@ -1,0 +1,6 @@
+import clsx, { type ClassValue } from "clsx"
+import { twMerge } from "tailwind-merge"
+
+export function cn(...args: ClassValue[]) {
+  return twMerge(clsx(...args))
+}
